+++ apollo-router-core/src/services/router_service.rs
@@ -1,4 +1,5 @@
-use crate::apq::APQ;
+use crate::apq::APQLayer;
+use crate::ensure_query_presence::EnsureQueryPresence;
 use crate::forbid_http_get_mutations::ForbidHttpGetMutationsLayer;
 use crate::services::execution_service::ExecutionService;
 use crate::{
@@ -7,17 +8,17 @@ use crate::{
     ResponseBody, RouterBridgeQueryPlanner, RouterRequest, RouterResponse, Schema, SubgraphRequest,
     SubgraphResponse,
 };
-use futures::future::BoxFuture;
-use futures::TryFutureExt;
+use futures::{future::BoxFuture, TryFutureExt};
+use http::StatusCode;
 use std::sync::Arc;
 use std::task::Poll;
 use tower::buffer::Buffer;
 use tower::util::{BoxCloneService, BoxService};
 use tower::{BoxError, ServiceBuilder, ServiceExt};
 use tower_service::Service;
-use tracing::instrument::WithSubscriber;
-use tracing::{Dispatch, Instrument};
 use typed_builder::TypedBuilder;
+
+static DEFAULT_BUFFER_SIZE: usize = 20_000;
 
 #[derive(TypedBuilder, Clone)]
 pub struct RouterService<QueryPlannerService, ExecutionService> {
@@ -79,29 +80,13 @@ where
         let schema = self.schema.clone();
         let query_cache = self.query_cache.clone();
 
-        let query = &req.context.request.body().query.as_ref();
-
-        if query.is_none() || query.expect("checked before; qed").is_empty() {
-            let res = plugin_utils::RouterResponse::builder()
-                .context(req.context.into())
-                .errors(vec![crate::Error {
-                    message: "Must provide query string.".to_string(),
-                    ..Default::default()
-                }])
-                .build()
-                .into();
-            return Box::pin(async move { Ok(res) });
-        };
-
-        let query = req
-            .context
-            .request
-            .body()
-            .query
-            .clone()
-            .expect("checked above; qed");
-
-        if let Some(response) = self.naive_introspection.get(query.as_str()) {
+        if let Some(response) =
+            self.naive_introspection.get(
+                req.context.request.body().query.as_ref().expect(
+                    "com.apollographql.ensure-query-is-present has checked this already; qed",
+                ),
+            )
+        {
             return Box::pin(async move {
                 Ok(RouterResponse {
                     response: http::Response::new(ResponseBody::GraphQL(response)).into(),
@@ -111,10 +96,10 @@ where
         }
 
         let fut = async move {
-            let body = req.context.request.body();
+            let context = req.context;
+            let body = context.request.body();
             let query = query_cache
-                .get_query(query.as_str())
-                .instrument(tracing::info_span!("query_parsing"))
+                .get_query(body.query.as_ref().expect("com.apollographql.ensure-query-is-present has checked this already; qed").as_str())
                 .await;
 
             if let Some(err) = query
@@ -123,16 +108,15 @@ where
             {
                 Ok(RouterResponse {
                     response: http::Response::new(ResponseBody::GraphQL(err)).into(),
-                    context: req.context.into(),
+                    context: context.into(),
                 })
             } else {
                 let operation_name = body.operation_name.clone();
                 let planned_query = planning
                     .call(QueryPlannerRequest {
-                        context: req.context.into(),
+                        context: context.into(),
                     })
                     .await?;
-
                 let mut response = execution
                     .call(ExecutionRequest {
                         query_plan: planned_query.query_plan,
@@ -145,7 +129,7 @@ where
                         query.format_response(
                             response.response.body_mut(),
                             operation_name.as_deref(),
-                            &schema,
+                            schema.api_schema(),
                         )
                     });
                 }
@@ -163,7 +147,7 @@ where
                     ..Default::default()
                 }])
                 .build()
-                .into())
+                .with_status(StatusCode::INTERNAL_SERVER_ERROR))
         });
 
         Box::pin(fut)
@@ -174,21 +158,19 @@ pub struct PluggableRouterServiceBuilder {
     schema: Arc<Schema>,
     buffer: usize,
     plugins: Vec<Box<dyn DynPlugin>>,
-    services: Vec<(
+    subgraph_services: Vec<(
         String,
         BoxService<SubgraphRequest, SubgraphResponse, BoxError>,
     )>,
-    dispatcher: Option<Dispatch>,
 }
 
 impl PluggableRouterServiceBuilder {
-    pub fn new(schema: Arc<Schema>, buffer: usize) -> Self {
+    pub fn new(schema: Arc<Schema>) -> Self {
         Self {
             schema,
-            buffer,
+            buffer: DEFAULT_BUFFER_SIZE,
             plugins: Default::default(),
-            services: Default::default(),
-            dispatcher: Default::default(),
+            subgraph_services: Default::default(),
         }
     }
 
@@ -202,11 +184,6 @@ impl PluggableRouterServiceBuilder {
 
     pub fn with_dyn_plugin(mut self, plugin: Box<dyn DynPlugin>) -> PluggableRouterServiceBuilder {
         self.plugins.push(plugin);
-        self
-    }
-
-    pub fn with_dispatcher(mut self, dispatcher: Dispatch) -> PluggableRouterServiceBuilder {
-        self.dispatcher = Some(dispatcher);
         self
     }
 
@@ -230,7 +207,8 @@ impl PluggableRouterServiceBuilder {
     where
         <S as Service<SubgraphRequest>>::Future: Send,
     {
-        self.services.push((name.to_string(), service.boxed()));
+        self.subgraph_services
+            .push((name.to_string(), service.boxed()));
         self
     }
 
@@ -254,31 +232,20 @@ impl PluggableRouterServiceBuilder {
             .unwrap_or(100);
 
         // QueryPlannerService takes an UnplannedRequest and outputs PlannedRequest
-        let (query_planner_service, query_worker) = Buffer::pair(
-            ServiceBuilder::new().service(
-                self.plugins.iter_mut().rev().fold(
-                    CachingQueryPlanner::new(
-                        RouterBridgeQueryPlanner::new(self.schema.clone()),
-                        plan_cache_limit,
-                    )
-                    .boxed(),
-                    |acc, e| e.query_planning_service(acc),
-                ),
-            ),
-            self.buffer,
-        );
-        tokio::spawn(
-            query_worker.with_subscriber(
-                self.dispatcher
-                    .as_ref()
-                    .expect("safe to assume dispatcher is some")
-                    .clone(),
+        let query_planner_service = ServiceBuilder::new().buffer(self.buffer).service(
+            self.plugins.iter_mut().rev().fold(
+                CachingQueryPlanner::new(
+                    RouterBridgeQueryPlanner::new(self.schema.clone()),
+                    plan_cache_limit,
+                )
+                .boxed(),
+                |acc, e| e.query_planning_service(acc),
             ),
         );
 
         // SubgraphService takes a SubgraphRequest and outputs a RouterResponse
         let subgraphs = self
-            .services
+            .subgraph_services
             .into_iter()
             .map(|(name, s)| {
                 let service = self
@@ -287,22 +254,15 @@ impl PluggableRouterServiceBuilder {
                     .rev()
                     .fold(s, |acc, e| e.subgraph_service(&name, acc));
 
-                let (service, worker) = Buffer::pair(service, self.buffer);
-                tokio::spawn(
-                    worker.with_subscriber(
-                        self.dispatcher
-                            .as_ref()
-                            .expect("safe to assume dispatcher is some")
-                            .clone(),
-                    ),
-                );
+                let service = ServiceBuilder::new().buffer(self.buffer).service(service);
 
                 (name.clone(), service)
             })
             .collect();
 
         // ExecutionService takes a PlannedRequest and outputs a RouterResponse
-        let (execution_service, execution_worker) = Buffer::pair(
+        // NB: Cannot use .buffer() here or the code won't compile...
+        let execution_service = Buffer::new(
             ServiceBuilder::new()
                 .layer(ForbidHttpGetMutationsLayer::default())
                 .service(
@@ -314,25 +274,19 @@ impl PluggableRouterServiceBuilder {
                             .boxed(),
                         |acc, e| e.execution_service(acc),
                     ),
-                ),
+                )
+                .boxed(),
             self.buffer,
-        );
-        tokio::spawn(
-            execution_worker.with_subscriber(
-                self.dispatcher
-                    .as_ref()
-                    .expect("safe to assume dispatcher is some")
-                    .clone(),
-            ),
         );
 
         let query_cache_limit = std::env::var("ROUTER_QUERY_CACHE_LIMIT")
             .ok()
             .and_then(|x| x.parse().ok())
             .unwrap_or(100);
-        let query_cache = Arc::new(QueryCache::new(query_cache_limit));
+        let query_cache = Arc::new(QueryCache::new(query_cache_limit, self.schema.clone()));
 
         // NaiveIntrospection instantiation can potentially block for some time
+        // We don't need to use the api schema here because on the deno side we always convert to API schema
         let naive_introspection = {
             let schema = self.schema.clone();
             tokio::task::spawn_blocking(move || NaiveIntrospection::from_schema(&schema))
@@ -360,29 +314,26 @@ impl PluggableRouterServiceBuilder {
         */
 
         // Router service takes a graphql::Request and outputs a graphql::Response
-        let (router_service, router_worker) = Buffer::pair(
-            ServiceBuilder::new().layer(APQ::default()).service(
-                self.plugins.iter_mut().rev().fold(
-                    RouterService::builder()
-                        .query_planner_service(query_planner_service)
-                        .query_execution_service(execution_service)
-                        .schema(self.schema)
-                        .query_cache(query_cache)
-                        .naive_introspection(naive_introspection)
-                        .build()
-                        .boxed(),
-                    |acc, e| e.router_service(acc),
-                ),
-            ),
+        // NB: Cannot use .buffer() here or the code won't compile...
+        let router_service = Buffer::new(
+            ServiceBuilder::new()
+                .layer(APQLayer::default())
+                .layer(EnsureQueryPresence::default())
+                .service(
+                    self.plugins.iter_mut().rev().fold(
+                        RouterService::builder()
+                            .query_planner_service(query_planner_service)
+                            .query_execution_service(execution_service)
+                            .schema(self.schema)
+                            .query_cache(query_cache)
+                            .naive_introspection(naive_introspection)
+                            .build()
+                            .boxed(),
+                        |acc, e| e.router_service(acc),
+                    ),
+                )
+                .boxed(),
             self.buffer,
-        );
-        tokio::spawn(
-            router_worker.with_subscriber(
-                self.dispatcher
-                    .as_ref()
-                    .expect("safe to assume dispatcher is some")
-                    .clone(),
-            ),
         );
 
         (router_service.boxed_clone(), self.plugins)