+++ fuzz/fuzz_targets/router.rs
@@ -1,9 +1,9 @@
 #![no_main]
-use apollo_router_core::Response;
+
 use libfuzzer_sys::fuzz_target;
 use log::debug;
 use router_fuzz::generate_valid_operation;
-use serde_json::json;
+use serde_json::{json, Value};
 use std::fs::OpenOptions;
 use std::io::Write;
 
@@ -24,19 +24,18 @@ fuzz_target!(|data: &[u8]| {
         .json(&json!({ "query": generated_operation }))
         .send()
         .unwrap()
-        .json::<Response>();
+        .json::<Value>();
     let gateway_response = http_client
         .post(GATEWAY_URL)
         .json(&json!({ "query": generated_operation }))
         .send()
         .unwrap()
-        .json::<Response>();
+        .json::<Value>();
 
     debug!("======= DOCUMENT =======");
     debug!("{}", generated_operation);
     debug!("========================");
     debug!("======= RESPONSE =======");
-    assert_eq!(router_response.is_ok(), gateway_response.is_ok());
     if router_response.is_ok() != gateway_response.is_ok() {
         let router_error = if let Err(err) = &router_response {
             Some(err)
@@ -74,9 +73,12 @@ fuzz_target!(|data: &[u8]| {
         );
         debug!("{errors}");
         file.write_all(errors.as_bytes()).unwrap();
+        file.flush().unwrap();
+
+        panic!()
     } else if router_response.is_ok() {
-        let router_response = router_response.unwrap();
-        let gateway_response = gateway_response.unwrap();
+        let router_response = serde_json::to_string_pretty(&router_response.unwrap()).unwrap();
+        let gateway_response = serde_json::to_string_pretty(&gateway_response.unwrap()).unwrap();
         if router_response != gateway_response {
             let mut file = OpenOptions::new()
                 .read(true)
@@ -94,16 +96,19 @@ fuzz_target!(|data: &[u8]| {
 {generated_operation}
 
 ====GATEWAY====
-{gateway_response:?}
+{gateway_response}
 
 ====ROUTER====
-{router_response:?}
+{router_response}
 
 
 "#
             );
             debug!("{errors}");
             file.write_all(errors.as_bytes()).unwrap();
+            file.flush().unwrap();
+
+            panic!();
         }
     }
     debug!("========================");